src/go2_ros2_toolbox_division_llm/pc_voice/requirements.txt
pyaudio
requests
pyyaml
openai-whisper
